Home Screen Shortcut Metadata
While signing up for the event, I noticed that the app didn't have an icon/preset name when adding it to my home screen for visiting it again later.

I looked through the source to see if I could find the original SVG of the icon, but it doesn't seem to be in the repo. This could be a placeholder until it's possible to do a proper export of that at 180x180, since I simply just re-scaled down the 512x512 favicon bilinearly.